requirements.txt
flask
requests
python-dateutil
APScheduler
reportlab
werkzeug
pymongo[srv]
python-dotenv
